fix: hook rename (1/2)
