+++ src/renderer/library/hook/useOutsideAlerter/index.ts
